Few changes to improve compatibility with 10.5
Most important is using --auth-root-authentication-method=normal which
changed the default value since 10.5 and while we don't use root acount
inside the container, we cannot use the new functionality that depends
on being a superuser after initialization of the datadir.

The replication test changes check of whether the slave has data from
the master.