Do not list the DB to be replicated
From some unknown reason this makes the user permissions not being replicated on 10.5 and further.

It seems that filtering on mysql name was not supposed to be working anyway according to
https://jira.mariadb.org/browse/MDEV-5704

And there were some changes done in 10.4 that might be related:
https://jira.mariadb.org/browse/MDEV-17658

Anyway, there is probably no reason to limit the databases replicated by default.