Replace syntax of chown from '.' to ':' for Dockerfiles C10S and RHEL10
----
Some older scripts may still use '.' in place of the ':' separator. POSIX 1003.1-2001
(see [Standards conformance](http://www.gnu.org/software/coreutils/manual/html_node/Standards-conformance.html))
does not require support for that, but for backward compatibility GNU chown supports '.' so long as no ambiguity results.
New scripts should avoid the use of '.' because it is not portable,
and because it has undesirable results if the entire owner'.'group happens to identify a user whose name contains '.'.
----

Signed-off-by: Petr "Stone" Hracek <phracek@redhat.com>